requirements.txt
numpy==1.20.1
pkg-resources==0.0.0
pyqtgraph==0.11.1
pyserial==3.5
numpy==1.20.1
pkg-resources==0.0.0
PyQt5==5.15.3
PyQt5-Qt==5.15.2
PyQt5-sip==12.8.1
pyqtgraph==0.11.1
pyserial==3.5
numpy==1.20.1
pkg-resources==0.0.0
PyQt5==5.15.3
PyQt5-Qt==5.15.2
PyQt5-sip==12.8.1
pyqtgraph==0.11.1
pyserial==3.5
